+++ src/main/java/com/trendist/issue_service/domain/issue/service/IssueService.java
@@ -22,7 +22,6 @@ import com.trendist.issue_service.domain.issue.domain.Issue;
 import com.trendist.issue_service.domain.issue.domain.IssueBookmark;
 import com.trendist.issue_service.domain.issue.domain.IssueDocument;
 import com.trendist.issue_service.domain.issue.dto.response.BookmarkResponse;
-import com.trendist.issue_service.domain.issue.dto.response.IssueGetAllBookmarkedResponse;
 import com.trendist.issue_service.domain.issue.dto.response.IssueGetAllResponse;
 import com.trendist.issue_service.domain.issue.dto.response.IssueGetResponse;
 import com.trendist.issue_service.domain.issue.dto.response.IssueSearchResponse;
@@ -118,15 +117,6 @@ public class IssueService {
 		}
 
 		return BookmarkResponse.of(bookmark, bookmarked);
-	}
-
-	public Page<IssueGetAllBookmarkedResponse> getAllIssuesBookmarked(int page) {
-		UUID userId = userServiceClient.getMyProfile("").getResult().id();
-
-		Pageable pageable = PageRequest.of(page, 10, Sort.by("issue.issueDate").descending());
-
-		return issueBookmarkRepository.findAllByUserId(userId, pageable)
-			.map(IssueGetAllBookmarkedResponse::from);
 	}
 
 	public Page<IssueSearchResponse> searchIssues(String keyword, int page) {
